test: story 파일을 ts로 변경하다
