--- src/components/NewCreditCardPageContent.stories.tsx
@@ -1,22 +1,0 @@
-import type { Meta } from '@storybook/react';
-import { HashRouter } from 'react-router-dom';
-import { NewCreditCardPageContent } from './NewCreditCardPageContent';
-
-const meta = {
-  title: 'NewCreditCardPageContent',
-  component: NewCreditCardPageContent,
-  tags: ["autodocs"],
-  decorators: [
-    (Story) => (
-      <HashRouter>
-        <Story />
-      </HashRouter>
-    ),
-  ],
-} satisfies Meta<typeof NewCreditCardPageContent>;
-
-export default meta;
-
-export const Default = () => {
-  return <NewCreditCardPageContent />;
-};
